feat(graph): ajoute la gestion des vues physiques et améliore la précision des positions
- Extension des vues disponibles dans les positions (ajout de "physique").
- Modification des types des coordonnées x, y, z de `int` à `float`.
- Suppression du paramètre `vue` dans la méthode `recalculate_positions` pour une plus grande flexibilité.

+++ app/schemas/GraphData.py
@@ -3,9 +3,9 @@ from typing import List, Optional, Dict
 from pydantic import BaseModel
 
 class Position(BaseModel):
-    x:Optional[int]
-    y:Optional[int]
-    z:Optional[int]
+    x: Optional[float]
+    y: Optional[float]
+    z: Optional[float]
 
 class Nodes(BaseModel):
     id: int
